Bug corrected in "insertBytes" Bugs corrected when a packet becomes empty and an ack is crafted Crafted ACK are sent on the second output of TCPOut Ack management improved

+++ elements/middlebox/tcpout.cc
@@ -35,6 +35,9 @@ Packet* TCPOut::processPacket(struct fcb* fcb, Packet* p)
     tcp_seq_t prevSeq = getSequenceNumber(packet);
     tcp_seq_t newSeq =  byteStreamMaintainer.mapSeq(prevSeq);
     bool seqModified = false;
+    bool ackModified = false;
+    tcp_seq_t prevAck = getAckNumber(packet);
+    tcp_seq_t prevLastAck = byteStreamMaintainer.getLastAck();
 
     if(prevSeq != newSeq)
     {
@@ -43,13 +46,27 @@ Packet* TCPOut::processPacket(struct fcb* fcb, Packet* p)
         seqModified = true;
     }
 
+    // Update the value of the last ACK sent
+    byteStreamMaintainer.setLastAck(prevAck);
+
+    // Ensure that the value of the ACK is not below the last ACKed position
+    /* This solves the following problem:
+     * - We ACK a packet manually for any reason
+     * - The "manual" ACK is lost
+     */
+    setAckNumber(packet, byteStreamMaintainer.getLastAck());
+
+    if(getAckNumber(packet) != prevAck)
+        ackModified = true;
+
     // Check if the packet has been modified
-    if(getAnnotationModification(packet) || seqModified)
+    if(getAnnotationModification(packet) || seqModified || ackModified)
     {
         // Check the length to see if bytes were added or removed
         uint16_t initialLength = IPElement::packetTotalLength(packet);
         uint16_t currentLength = (uint16_t)packet->length() - IPElement::getIPHeaderOffset(packet);
         int offsetModification = -(initialLength - currentLength);
+        uint32_t prevPayloadSize = getPayloadLength(packet) - offsetModification;
 
         // Update the "total length" field in the IP header (required to compute the tcp checksum as it is in the pseudo hdr)
         IPElement::setPacketTotalLength(packet, initialLength + offsetModification);
@@ -60,7 +77,6 @@ Packet* TCPOut::processPacket(struct fcb* fcb, Packet* p)
             // We know that the packet has been modified and its size has changed
             modList->commit(fcb->tcp_common->maintainers[getFlowDirection()]);
 
-            // TODO: Rework that part
             // Check if the full packet content has been removed
             if(getPayloadLength(packet) == 0)
             {
@@ -68,20 +84,42 @@ Packet* TCPOut::processPacket(struct fcb* fcb, Packet* p)
                 uint32_t daddr = IPElement::getSourceAddress(packet);
                 uint16_t sport = getDestinationPort(packet);
                 uint16_t dport = getSourcePort(packet);
+                // The SEQ value is the initial ACK value in the packet sent
+                // by the source.
+                // To get it, we take the current one (which has been corrected)
+                // and we map it as if it was a sequence number to get back
+                // the initial value
                 tcp_seq_t seq = getAckNumber(packet);
-                uint8_t winSize = getWindowSize(packet);
-                tcp_seq_t ack = prevSeq + getPayloadLength(packet);
+                seq = fcb->tcp_common->maintainers[getOppositeFlowDirection()].mapSeq(seq);
+                uint16_t winSize = getWindowSize(packet);
+                // The ACK is the sequence number sent by the source
+                // to which we add the old size of the payload to acknowledge it
+                tcp_seq_t ack = prevSeq + prevPayloadSize;
 
-                // The packet is now empty, we discard it and send an ACK directly to the source
-                click_chatter("Empty packet. Sending an ACK! (%u - %u - %d)", saddr, daddr, offsetModification);
+                if(isFinOrSyn(packet))
+                    ack++;
 
-                Packet* forged = forgePacket(saddr, daddr, sport, dport, seq, ack, winSize, TH_ACK);
-                packet->kill();
+                // Craft and send the ack
+                sendAck(saddr, daddr, sport, dport, seq, ack, winSize);
 
-                // Send it
-                inElement->getReturnElement()->getOutElement()->push(0, forged);
+                // Update the number of the last ack sent for the other side
+                fcb->tcp_common->maintainers[getOppositeFlowDirection()].setLastAck(ack);
 
-                return NULL;
+                // Even if the packet is empty it can still contain relevant
+                // information (significant ACK value or another flag)
+                if(isJustAnAck(packet))
+                {
+                    // Check if the ACK of the packet was significant or not
+                    if(prevAck <= prevLastAck)
+                    {
+                        // If this is not the case, drop the packet as it
+                        // does not contain any relevant information
+                        // (And anyway it would be considered as a duplicate ACK)
+                        click_chatter("Empty packet dropped");
+                        packet->kill();
+                        return NULL;
+                    }
+                }
             }
         }
 
@@ -90,6 +128,24 @@ Packet* TCPOut::processPacket(struct fcb* fcb, Packet* p)
     }
 
     return packet;
+}
+
+void TCPOut::sendAck(uint32_t saddr, uint32_t daddr, uint16_t sport, uint16_t dport, tcp_seq_t seq, tcp_seq_t ack, uint16_t winSize)
+{
+    if(noutputs() < 2)
+    {
+        click_chatter("Warning: trying to send an ack on a TCPOut with only 1 output");
+        return;
+    }
+
+    // The packet is now empty, we discard it and send an ACK directly to the source
+    click_chatter("Sending an ACK! (%u -> %u : %u)", saddr, daddr, ack);
+
+    // Craft the packet
+    Packet* forged = forgePacket(saddr, daddr, sport, dport, seq, ack, winSize, TH_ACK);
+
+    //Send it on the second output
+    push(1, forged);
 }
 
 void TCPOut::setInElement(TCPIn* inElement)
